refactor: 코드 컨벤션에 맞게 패키지 구조 개편
- com.mio.domain.{domain} → com.mio.{domain}.domain (엔티티는 domain 서브패키지)
- com.mio.ai.memory → com.mio.ai.domain
- 모든 도메인이 루트 레벨 패키지로 이동
- 교차 도메인 import 경로 전면 업데이트
