--- src/main/java/com/mio/domain/dailytest/DailyTestResponse.java
@@ -1,43 +1,0 @@
-package com.mio.domain.dailytest;
-
-import com.mio.domain.user.User;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.OffsetDateTime;
-import java.util.UUID;
-
-@Entity
-@Table(name = "daily_test_responses")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Builder
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class DailyTestResponse {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "daily_test_id", nullable = false)
-    private DailyTest dailyTest;
-
-    @Column(name = "answers", nullable = false, columnDefinition = "jsonb")
-    private String answers;
-
-    @Column(name = "result_summary")
-    private String resultSummary;
-
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private OffsetDateTime createdAt;
-
-    @PrePersist
-    protected void onCreate() {
-        createdAt = OffsetDateTime.now();
-    }
-}
