+++ src/main/java/com/mio/user/domain/UserConsent.java
@@ -1,0 +1,41 @@
+package com.mio.user.domain;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.OffsetDateTime;
+import java.util.UUID;
+
+@Entity
+@Table(name = "user_consents")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class UserConsent {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @Column(name = "consent_type", nullable = false)
+    private String consentType;
+
+    @Column(name = "agreed", nullable = false)
+    private boolean agreed;
+
+    @Column(name = "agreed_at", nullable = false)
+    private OffsetDateTime agreedAt;
+
+    @Column(name = "version", nullable = false)
+    private String version;
+
+    @PrePersist
+    protected void onCreate() {
+        agreedAt = OffsetDateTime.now();
+    }
+}
